小更新，如下：

+++ src/main/java/com/rbs/project/controller/SeminarController.java
@@ -4,13 +4,15 @@ import com.rbs.project.exception.MyException;
 import com.rbs.project.pojo.dto.CreateSeminarDTO;
 import com.rbs.project.pojo.dto.UpdateSeminarDTO;
 import com.rbs.project.pojo.entity.*;
+import com.rbs.project.pojo.vo.QuestionInfoVO;
 import com.rbs.project.service.*;
 import com.rbs.project.utils.JsonUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -36,7 +38,7 @@ public class SeminarController {
      */
     @PostMapping("")
     @ResponseBody
-    public ResponseEntity<Long> createSeminar(@RequestBody CreateSeminarDTO createSeminarDTO) throws MyException {
+    public ResponseEntity<Long> createSeminar(@RequestBody CreateSeminarDTO createSeminarDTO) throws Exception {
         //初始化新增讨论课id
         long createSeminarId = -1;
         Seminar seminar = new Seminar();
@@ -70,8 +72,14 @@ public class SeminarController {
      */
     @GetMapping("/{seminarId}/")
     @ResponseBody
-    public ResponseEntity<Map<String, Object>> getSeminarById(@PathVariable("seminarId") long seminarId, @RequestParam("cClassId") long cClassId) throws MyException {
-        Seminar seminar = seminarService.getSeminarById(seminarId);
+    public ResponseEntity<Map<String,Object>> getSeminarById(@PathVariable("seminarId")long seminarId,@RequestParam("cClassId") long cClassId) throws MyException{
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        if((Long)cClassId==null){
+            throw new MyException("classId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        Seminar seminar=seminarService.getSeminarById(seminarId);
         //转换格式
         Map<String, Object> seminarView = new HashMap<>();
         seminarView.put("courseId", seminar.getCourseId());
@@ -94,7 +102,10 @@ public class SeminarController {
      */
     @DeleteMapping("/{seminarId}")
     @ResponseBody
-    public ResponseEntity<Boolean> deleteSeminarById(@PathVariable("seminarId") long seminarId) throws MyException {
+    public ResponseEntity<Boolean> deleteSeminarById(@PathVariable("seminarId") long seminarId) throws MyException{
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
         return ResponseEntity.ok().body(seminarService.removeSeminarById(seminarId));
     }
 
@@ -106,7 +117,10 @@ public class SeminarController {
      */
     @PutMapping("/{seminarId}")
     @ResponseBody
-    public ResponseEntity<Boolean> updateSeminarById(@PathVariable("seminarId") long seminarId, @RequestBody UpdateSeminarDTO updateSeminarDTO) throws MyException {
+    public ResponseEntity<Boolean> updateSeminarById(@PathVariable("seminarId") long seminarId, @RequestBody UpdateSeminarDTO updateSeminarDTO) throws Exception {
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
         //DTO转Entity
         Seminar seminar = new Seminar();
         seminar.setId(seminarId);
@@ -130,10 +144,72 @@ public class SeminarController {
     @ResponseBody
     public ResponseEntity<Boolean> updateCClassSeminar(@PathVariable("seminarId") long seminarId,
                                                        @PathVariable("classId") long cClassId,
-                                                       @RequestBody CClassSeminar cClassSeminar) throws MyException {
+                                                       @RequestBody CClassSeminar cClassSeminar) throws MyException{
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        if((Long)cClassId==null){
+            throw new MyException("classId不能为空",MyException.ID_FORMAT_ERROR);
+        }
         cClassSeminar.setcClassId(cClassId);
         cClassSeminar.setSeminarId(seminarId);
         return ResponseEntity.ok().body(cClassSeminarService.updateCClassSeminar(cClassSeminar));
     }
 
+    /**
+     * Description: 查看班级下讨论课的所有提问
+     * ！！！！应有websocket！！！！！
+     *
+     * @Author: WinstonDeng
+     * @Date: 16:05 2018/12/22
+     */
+    @GetMapping("/{seminarId}/class/{classId}/question")
+    @ResponseBody
+    public ResponseEntity<List<QuestionInfoVO>> listAllQuestions(@PathVariable("seminarId")long seminarId, @PathVariable("classId") long cClassId) throws MyException{
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        if((Long)cClassId==null){
+            throw new MyException("classId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        List<Question> questions=cClassSeminarService.listAllQuestionsByCClassIdAndSeminarId(cClassId,seminarId);
+        List<QuestionInfoVO> questionInfoVOS=new ArrayList<>();
+        for(Question question:questions){
+            questionInfoVOS.add(new QuestionInfoVO(question));
+        }
+        return ResponseEntity.ok().body(questionInfoVOS);
+    }
+
+    /**
+     * Description: 修改班级下讨论课的提问
+     *   1.选择提问
+     *   2.打分
+     *   ！！！！！websocket通知！！！！！！
+     * @Author: WinstonDeng
+     * @Date: 16:40 2018/12/22
+     */
+    @PutMapping("/{seminarId}/class/{classId}/question")
+    @ResponseBody
+    public ResponseEntity<Boolean> updateQuestion(@PathVariable("seminarId")long seminarId,
+                                                  @PathVariable("classId") long cClassId,
+                                                  @RequestBody QuestionInfoVO questionInfoVO) throws MyException{
+        if((Long)seminarId==null){
+            throw new MyException("seminarId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        if((Long)cClassId==null){
+            throw new MyException("classId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        Question question=new Question();
+        question.setScore(questionInfoVO.getScore());
+        if(questionInfoVO.getSelected()==true){
+            question.setSelected(1);
+        }else if(questionInfoVO.getSelected()==false){
+            question.setSelected(0);
+        }
+        if(questionInfoVO.getQuestionId()==null){
+           throw new MyException("questionId不能为空",MyException.ID_FORMAT_ERROR);
+        }
+        question.setId(questionInfoVO.getQuestionId());
+        return ResponseEntity.ok().body(cClassSeminarService.updateQuestion(question));
+    }
 }
